Create Practice 17.dart
For this exercise, we will keep track of when our friend’s birthdays are, and be able to find that information based on their name.

Create a dictionary (in your file) of names and birthdays. When you run your program it should ask the user to enter a name, and return the birthday of that person back to them. The interaction should look something like this: